OF-1097: Reduce complexity of AuthProvider AuthProvider has quite some complexity that exists solely for XEP-0078. This commit removes most of that, by replacing the generic checks for digest and plain support with a non-generic implementation, that will work for any auth provider that supports password retrieval.
