+++ src/java/org/jivesoftware/openfire/auth/POP3AuthProvider.java
@@ -220,23 +220,6 @@ public class POP3AuthProvider implements AuthProvider {
     }
 
     @Override
-    public void authenticate(String username, String token, String digest)
-            throws UnauthorizedException
-    {
-        throw new UnauthorizedException("Digest authentication not supported.");
-    }
-
-    @Override
-    public boolean isPlainSupported() {
-        return true;
-    }
-
-    @Override
-    public boolean isDigestSupported() {
-        return false;
-    }
-
-    @Override
     public String getPassword(String username)
             throws UserNotFoundException, UnsupportedOperationException
     {
